fix: move the store (redux) component into the infrastructure section
